+++ pokedex/src/components/HomePage.js
@@ -18,17 +18,17 @@ import { goToDetails, goToPokedex } from "./router/Coordinator";
 import { useHistory } from "react-router-dom";
 import Logomarca from "../img/pokemon-logo.png"
 import GlobalStateContext from "../global/GlobalStateContext";
+import CardPokemon from "./CardPokemon";
 
 
 
 function HomePage() {
   const history = useHistory();
-  const { pokeImage, setPokeImage, getPokeImage, pokemons, setPokemons, getPokemons, pokedex, setPodedex, pokeDetails, setPokeDetails } = useContext(GlobalStateContext);
+  const { urlImage, setUrlImage, pokeImage, setPokeImage, getPokeImage, pokemons, setPokemons, getPokemons, pokedex, setPodedex, pokeDetails, setPokeDetails } = useContext(GlobalStateContext);
     
   
   useEffect(() => {
     getPokemons();
-    getPokeImage();
   }, []);
 
   return <MainContainer>
@@ -42,16 +42,7 @@ function HomePage() {
             </HeaderContainer>
             <BodyContainerList>
             {pokemons.map((poke) => {
-              return <CardContainer>
-                      <ImgPokemon src= {pokeImage}></ImgPokemon>
-                      <div>
-                        <TextTitle>{poke.name}</TextTitle>
-                      </div>
-                      <ButtonContainer>
-                        <YellowButton onClick={() => goToPokedex(history)}>ADICIONAR À POKEDEX</YellowButton>
-                        <YellowButton onClick={() => goToDetails(history)}>VER DETALHES</YellowButton>
-                      </ButtonContainer>
-                    </CardContainer> 
+              return <CardPokemon name={poke.name} url={poke.url} />
             })}
             </BodyContainerList>
         </MainContainer>;
